creates team athletes modal list view
